Hyojin-Kim/Semantic_Segmentation/requirements.txt
absl-py==0.12.0
adamp==0.3.0
addict==2.4.0
albumentations==0.5.2
antlr4-python3-runtime==4.8
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
colorful==0.5.4
crfseg==1.0.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
# Editable install with no version control (detectron2==0.4)
-e /opt/conda/envs/second/lib/python3.7/site-packages
dill==0.3.3
easydict==1.9
efficientnet-pytorch==0.6.3
future==0.18.2
fvcore==0.1.5.post20210423
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.0.1
iopath==0.1.8
kiwisolver==1.3.1
madgrad==1.1
Markdown==3.3.4
matplotlib==3.4.1
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853974840/work
mkl-service==2.3.0
mmcv-full==1.3.1
-e git+https://github.com/open-mmlab/mmsegmentation.git@d568d06e75dc815309fd9f7fb9b422498c9ad213#egg=mmsegmentation
multiprocess==0.70.11.1
munch==2.5.0
networkx==2.5.1
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.0.dev22
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.2.4
Pillow @ file:///tmp/build/80754af9/pillow_1617386154241/work
portalocker==2.3.0
pretrainedmodels==0.7.4
prettyprinter==0.18.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
pydot==1.4.2
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.3
segmentation-models-pytorch==0.1.3
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1605205313296/work
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminaltables==3.1.0
tifffile==2021.4.8
timm==0.3.2
torch==1.6.0+cu101
torchsummary==1.5.1
torchvision==0.7.0+cu101
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
yacs==0.1.8
yapf==0.31.0
zipp==3.4.1
